% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Loads the Break Free From Plastic dataset.}
\usage{
load_data()
}
\value{
A tibble containing plastic waste data.
}
\description{
Loads the Break Free From Plastic dataset.
}
\examples{
load_data()
}
