% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_fertility_country.R
\name{validate_fertility_data}
\alias{validate_fertility_data}
\title{Validate Fertility Data}
\usage{
validate_fertility_data(combined)
}
\arguments{
\item{combined}{A data frame.}
}
\value{
The original data frame if validation passes.
}
\description{
Checks that the input data has the columns needed for the fertility animation.
}
